+++ backend/src/main/java/no/ntnu/database/controller/CategoryController.java
@@ -1,9 +1,9 @@
-package no.ntnu.database.controllers;
+package no.ntnu.database.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import no.ntnu.database.model.Category;
-import no.ntnu.database.services.CategoryService;
+import no.ntnu.database.service.CategoryService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
