+++ backend/src/main/java/no/ntnu/database/controller/CourseProviderController.java
@@ -1,11 +1,11 @@
-package no.ntnu.database.controllers;
+package no.ntnu.database.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.Optional;
 import no.ntnu.database.model.CourseProvider;
-import no.ntnu.database.services.CourseProviderService;
+import no.ntnu.database.service.CourseProviderService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
