+++ backend/src/main/java/no/ntnu/database/jwt/AccessUserService.java
@@ -3,7 +3,7 @@ package no.ntnu.database.jwt;
 
 import java.util.Optional;
 import no.ntnu.database.model.User;
-import no.ntnu.database.repositories.UserRepository;
+import no.ntnu.database.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
