+++ backend/src/main/java/no/ntnu/database/repository/ImageRepository.java
@@ -1,4 +1,4 @@
-package no.ntnu.database.repositories;
+package no.ntnu.database.repository;
 
 import no.ntnu.database.model.Image;
 import org.springframework.data.repository.CrudRepository;
