+++ backend/src/main/java/no/ntnu/database/service/CourseProviderService.java
@@ -1,8 +1,8 @@
-package no.ntnu.database.services;
+package no.ntnu.database.service;
 
 import java.util.Optional;
 import no.ntnu.database.model.CourseProvider;
-import no.ntnu.database.repositories.CourseProviderRepository;
+import no.ntnu.database.repository.CourseProviderRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
